+++ packages/eyes-storybook/src/getIframeUrl.js
@@ -3,11 +3,15 @@ const {URL} = require('url');
 
 function getIframeUrl(baseUrl) {
   const {origin, pathname} = new URL(baseUrl);
-  let baseUrlFixed = `${origin}${pathname}`;
+
+  let baseUrlFixed = `${origin}${pathname.replace(/\/[^\/]+\.html/, '')}`;
+
   if (!/\/$/.test(baseUrlFixed)) {
     baseUrlFixed += '/';
   }
-  return new URL(`iframe.html?eyes-storybook=true`, baseUrlFixed).href;
+
+  const url = new URL(`iframe.html?eyes-storybook=true`, baseUrlFixed);
+  return url.href;
 }
 
 module.exports = getIframeUrl;